refactor: [#1211] move tests to AnnounceHandler
